backend/requirements.txt
psutil>=5.9.0
fastapi
psycopg2-binary
bcrypt